feat(full-node): moved full node routes to their own subdir
