+++ package.json
@@ -11,7 +11,6 @@
     "start": "node ./src/app.js",
     "dev": "nodemon ./dist/app.js",
     "test": "npm run test-v3",
-    "test-v2": "export NETWORK=testnet && nyc --reporter=text mocha --timeout 25000 test/v2/",
     "test-v3": "export NETWORK=mainnet && nyc --reporter=text mocha --timeout 25000 test/v3/",
     "test:integration": "mocha test/v3/integration",
     "coverage": "nyc report --reporter=text-lcov | coveralls",
