+++ src/app.js
@@ -22,11 +22,11 @@ const jwtAuth = require("./middleware/jwt-auth")
 
 // v3
 const healthCheckV3 = require("./routes/v3/health-check")
-const blockchainV3 = require("./routes/v3/blockchain")
-const controlV3 = require("./routes/v3/control")
-const miningV3 = require("./routes/v3/mining")
-const networkV3 = require("./routes/v3/network")
-const rawtransactionsV3 = require("./routes/v3/rawtransactions")
+const blockchainV3 = require("./routes/v3/full-node/blockchain")
+const controlV3 = require("./routes/v3/full-node/control")
+const miningV3 = require("./routes/v3/full-node/mining")
+const networkV3 = require("./routes/v3/full-node/network")
+const rawtransactionsV3 = require("./routes/v3/full-node/rawtransactions")
 const utilV3 = require("./routes/v3/util")
 const slpV3 = require("./routes/v3/slp")
 const xpubV3 = require("./routes/v3/xpub")
