+++ src/routes/v3/full-node/network.js
