--- test/v2/integration/README.md
@@ -1,7 +1,0 @@
-This directory contains integration tests. To run them requires that
-a local copy of rest.bitcoin.com is running. The tests in this directory are
-exactly the same calls that end users would use for their app to interact
-directly with rest.bitcoin.com.
-
-At present, this directory focuses on raw-transations. But more integration tests
-will be added.
