--- test/v2/mocks/util-mocks.js
@@ -1,18 +1,0 @@
-/*
-  This library contains mocking data for running unit tests on the address route.
-*/
-
-"use strict"
-
-const mockAddress = {
-  isvalid: true,
-  address: "bchtest:qqqk4y6lsl5da64sg5qc3xezmplyu5kmpyz2ysaa5y",
-  scriptPubKey: "76a914016a935f87e8deeab04501889b22d87e4e52db0988ac",
-  ismine: false,
-  iswatchonly: false,
-  isscript: false
-}
-
-module.exports = {
-  mockAddress
-}
