+++ test/v3/raw-transactions.js
@@ -14,7 +14,7 @@
 
 const chai = require("chai")
 const assert = chai.assert
-const rawtransactions = require("../../src/routes/v3/rawtransactions")
+const rawtransactions = require("../../src/routes/v3/full-node/rawtransactions")
 const nock = require("nock") // HTTP mocking
 
 let originalEnvVars // Used during transition from integration to unit tests.
